add ccx discover interface
ccx exits in most AMD/hygon servers. it is defined as a cpu range those
share the same L3 cache. ccx is a subregion of a numa region. When we
build the resource pool, ccx can be better a leaf node than numa node.

This commit add hw discover interface for ccx alongside with nodes, cpus
and so on.

Signed-off-by: Jianyong Wu <jianyong.wu@outlook.com>