topology: fix max cpu allocation crash issue
For example, we have one k8s node with 1 socket and with 2 numa node
inside. Each of numa node has 16 cpus. There is one cpu reserved for
kube-system, so the most allocatble cpu number is 31. When a pod deployed
to this node requests 31 cpus, the sort algrithm of those three resource
pools list as below:
1. shared score for numa node 1, short as score1, is 16000 - 31000 = -15000,;
   shared score for socket 0, short as score2, is 31000 - 31000 = 0
   but as the first entry of sort algrithm, they will be tie, as the
   win condition is score1 > 0 or score2 > 0.

omit 2, 3, 4, when get into 5:
5. numa node 0 win for the lower depth in the pool tree.
in the end numa node 1 is assigned to the pod which is totally wrong as it has
no enough cpu num as requested by the pod then cri-rm will crash.

The root cause derives from the first sort algrithm. it should be:
(shared2 < 0 && shared1 > 0
not
(shared2 <= 0 && shared1 > 0

Signed-off-by: Jianyong Wu <jianyong.wu@outlook.com>