test: Enable test cases in pools_test.go
Some test cases inside pools_test.go are skipped as it needs a copy of
sysfs in real hygon machine.
To overcome this, here we introduce a copy of sysfs from a hygon #1
server. Also make some change inside test file and code file to adapt
it.
After that, unit tests coverage promoted from 49.8% to 61.8% that catch
up with the original one which is 60.1%.

Change-Id: I62d9c20287af2faa03a3e4befbd61cc234ee46f7