bugfix: resource resolve doesn't work for numa+ccx
There is a bug in the last patch which introduces possible cpu in
compareScores. It can't allocate numa + ccx for request cpu number
between one numa cap and one numa plus one ccx cap. The possible cpu
only used for numa and ccx should not be affected. So, exclude ccx for
that rule.

Signed-off-by: Jianyong Wu <wujianyong@hygon.cn>
Change-Id: Ie7c952477f29ca24a1916764ac3a795a054147a8